+++ app/general/App.js
@@ -1,17 +1,25 @@
-import React from 'react';
+import React, { Suspense } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { Router } from './Router';
 import { AlertProvider, Alert } from 'meteor/quave:alert-react-tailwind';
+import { Loading } from '../components/Loading';
+import { PageWithHeader } from '../layouts/PageWithHeader';
 
 export const App = () => (
   <BrowserRouter>
-    <AlertProvider>
-      <div className="h-full bg-indigo-50">
-        <Alert />
-        <div className="mx-auto max-w-7xl py-12 px-4 sm:px-6 lg:flex lg:items-center lg:justify-between lg:py-24 lg:px-8">
+    <Suspense
+      fallback={
+        <PageWithHeader>
+          <Loading name="suspense" />
+        </PageWithHeader>
+      }
+    >
+      <AlertProvider>
+        <div className="h-full bg-indigo-50">
+          <Alert />
           <Router />
         </div>
-      </div>
-    </AlertProvider>
+      </AlertProvider>
+    </Suspense>
   </BrowserRouter>
 );