+++ app/general/RoutePaths.js
@@ -1,4 +1,5 @@
 export const RoutePaths = {
   HOME: '/',
   ACCESS: '/access',
+  PRIVATE: '/private',
 };