+++ app/layouts/AnonymousLayout.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { ConditionalLayout } from './ConditionalLayout';
+
+export const AnonymousLayout = ({ children }) => (
+  <ConditionalLayout onlyAnonymous>{children}</ConditionalLayout>
+);