+++ app/layouts/LoggedLayout.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { ConditionalLayout } from './ConditionalLayout';
+
+export const LoggedLayout = ({ children }) => (
+  <ConditionalLayout onlyLogged>{children}</ConditionalLayout>
+);