+++ app/layouts/PageWithHeader.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const PageWithHeader = ({ children }) => (
+  <div className="flex h-full flex-col">{children}</div>
+);