+++ app/layouts/PublicLayout.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { ConditionalLayout } from './ConditionalLayout';
+
+export const PublicLayout = ({ children }) => (
+  <ConditionalLayout>{children}</ConditionalLayout>
+);